Initial commit of bbengine
Very unstructured, but somewhat functional.  Naive CONFI algorithm is
presented, but there are a few optimizations we still need.

The following few commits should aim to:

1) implement the “Shape” criterion properly
2) separate out CONFI code from the main library since it’s WAY too
specialized
3) consider making cards/hands easier to work with instead of just
representing them as a string.  it’d be nice at a high level to ask if
a suit is “as good as” Qxxx, for example.
4) more sanity checks, i.e. we’re have 13 cards per hand!
5) add a README.md
6) fix pylint output - it actually has some good suggestions like
“don’t name this function ‘do’”
7) add a debugging system so I don’t have to keep commenting and
uncommenting large swaths of print statements during testing
8) Clean up Auctioneer/bidder interface